+++ backend/canisters/local_user_index/api/src/updates/c2c_group_canister.rs
@@ -1,10 +1,13 @@
 use crate::GroupEvent;
 use serde::{Deserialize, Serialize};
-use types::{IdempotentEnvelope, UnitResult};
+use types::IdempotentEnvelope;
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Args {
     pub events: Vec<IdempotentEnvelope<GroupEvent>>,
 }
 
-pub type Response = UnitResult;
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub enum Response {
+    Success,
+}